backend/requirements.txt
Django==5.2.1
asgiref==3.8.1
sqlparse==0.5.3
djangorestframework==3.16.0
django-filter==25.1
djoser==2.3.1
psycopg2-binary==2.9.10
Pillow==11.2.1
gunicorn==23.0.0
